+++ packages/components/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@synergy-design-system/components-v2.18.6](https://github.com/synergy-design-system/synergy-design-system/compare/components/2.18.5...components/2.18.6) (2025-01-20)
+
+
+### Bug Fixes
+
+* 🐛 syn-range Various issues for syn-range ([#742](https://github.com/synergy-design-system/synergy-design-system/issues/742)) ([d6dc350](https://github.com/synergy-design-system/synergy-design-system/commit/d6dc3505576aa65cadb7c78f7a4a44deeffb3e4a))
+
 # [@synergy-design-system/components-v2.18.5](https://github.com/synergy-design-system/synergy-design-system/compare/components/2.18.4...components/2.18.5) (2025-01-20)